+++ src/main/java/com/mallang/statistics/api/query/StatisticCondition.java
@@ -1,6 +1,5 @@
 package com.mallang.statistics.api.query;
 
-import com.mallang.statistics.api.query.support.PeriodType;
 import java.time.LocalDate;
 
 public record StatisticCondition(
@@ -8,5 +7,4 @@ public record StatisticCondition(
         LocalDate startDayInclude,
         LocalDate lastDayInclude
 ) {
-
 }
